ansible: new pattern for `ops`
Let's have the top-level `bin` as clean as possible. We don't want the
user to understand the whole details of the deployment, so ideally just
the deploy as a "public" bin is enough.

Moreover, let's ensure it reads the dotfiles, as well as download
dependencies from ansible before starting the provison process.